requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
black==23.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
filelock==3.12.0
flake8==6.0.0
Flask==2.2.3
Flask-Cors==3.0.10
frozenlist==1.3.3
fsspec==2023.4.0
huggingface-hub==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mccabe==0.7.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
numpy==1.24.3
openai==0.27.4
packaging==23.1
pathspec==0.11.1
platformdirs==3.2.0
pycodestyle==2.10.0
pyflakes==3.0.1
python-dotenv==1.0.0
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
six==1.16.0
sympy==1.11.1
tokenizers==0.13.3
tomli==2.0.1
torch==2.0.0
tqdm==4.65.0
transformers==4.28.1
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
yarl==1.9.1
